List of stats

+++ BasketAPI/Model/Stats.swift
@@ -13,24 +13,24 @@ struct Stats: Codable, Identifiable{
     let player_id: Int
     let season: Int
     let min: String
-    let fgm: Int
-    let fga: Int
-    let fg3m: Int
-    let fg3a: Int
-    let ftm: Int
-    let fta: Int
-    let oreb: Int
-    let dreb: Int
-    let reb: Int
-    let ast: Int
-    let stl: Int
-    let blk: Int
-    let turnover: Int
-    let pf: Int
-    let pts: Int
-    let fg_pct: Int
-    let fg3_pct: Int
-    let ft_pct: Int
+    let fgm: Double
+    let fga: Double
+    let fg3m: Double
+    let fg3a: Double
+    let ftm: Double
+    let fta: Double
+    let oreb: Double
+    let dreb: Double
+    let reb: Double
+    let ast: Double
+    let stl: Double
+    let blk: Double
+    let turnover: Double
+    let pf: Double
+    let pts: Double
+    let fg_pct: Double
+    let fg3_pct: Double
+    let ft_pct: Double
     
 }
